Refactor lint subparser and hook tests
