+++ numpydoc/tests/hooks/test_validate_hook.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import pytest
 
-from numpydoc.hooks.validate_docstrings import get_parser, run_hook
+from numpydoc.hooks.validate_docstrings import run_hook
 
 
 @pytest.fixture
@@ -18,11 +18,6 @@ def example_module(request):
         / "example_module.py"
     )
     return str(fullpath.relative_to(request.config.rootdir))
-
-
-def hook_main(args):
-    parser = get_parser()
-    return run_hook(**vars(parser.parse_args(args)))
 
 
 @pytest.mark.parametrize("config", [None, "fake_dir"])
@@ -77,11 +72,7 @@ def test_validate_hook(example_module, config, capsys):
         """
     )
 
-    args = [example_module]
-    if config:
-        args.append(f"--{config=}")
-
-    return_code = hook_main(args)
+    return_code = run_hook([example_module], config=config)
     assert return_code == 1
     assert capsys.readouterr().err.rstrip() == expected
 
@@ -114,7 +105,7 @@ def test_validate_hook_with_ignore(example_module, capsys):
         """
     )
 
-    return_code = hook_main([example_module, "--ignore", "ES01", "SA01", "EX01"])
+    return_code = run_hook([example_module], ignore=["ES01", "SA01", "EX01"])
 
     assert return_code == 1
     assert capsys.readouterr().err.rstrip() == expected
@@ -163,7 +154,7 @@ def test_validate_hook_with_toml_config(example_module, tmp_path, capsys):
         """
     )
 
-    return_code = hook_main([example_module, "--config", str(tmp_path)])
+    return_code = run_hook([example_module], config=tmp_path)
     assert return_code == 1
     assert capsys.readouterr().err.rstrip() == expected
 
@@ -202,21 +193,9 @@ def test_validate_hook_with_setup_cfg(example_module, tmp_path, capsys):
         """
     )
 
-    return_code = hook_main([example_module, "--config", str(tmp_path)])
+    return_code = run_hook([example_module], config=tmp_path)
     assert return_code == 1
     assert capsys.readouterr().err.rstrip() == expected
-
-
-def test_validate_hook_help(capsys):
-    """Test that help section is displaying."""
-
-    with pytest.raises(SystemExit):
-        return_code = hook_main(["--help"])
-        assert return_code == 0
-
-    out = capsys.readouterr().out
-    assert "--ignore" in out
-    assert "--config" in out
 
 
 def test_validate_hook_exclude_option_pyproject(example_module, tmp_path, capsys):
@@ -261,7 +240,7 @@ def test_validate_hook_exclude_option_pyproject(example_module, tmp_path, capsys
         """
     )
 
-    return_code = hook_main([example_module, "--config", str(tmp_path)])
+    return_code = run_hook([example_module], config=tmp_path)
     assert return_code == 1
     assert capsys.readouterr().err.rstrip() == expected
 
@@ -298,6 +277,6 @@ def test_validate_hook_exclude_option_setup_cfg(example_module, tmp_path, capsys
         """
     )
 
-    return_code = hook_main([example_module, "--config", str(tmp_path)])
+    return_code = run_hook([example_module], config=tmp_path)
     assert return_code == 1
     assert capsys.readouterr().err.rstrip() == expected
